pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "heavyedge-features"
version = "0.0.0"
authors = [
    {name = "Jisoo Song", email = "jeesoo9595@snu.ac.kr"}
]
description = "Package to quantify edge profile shape features"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = [
    "LICENSE",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering",
    "Operating System :: OS Independent",
]
dependencies = [
    "heavyedge",
    "cvxpy",
]

[project.urls]
homepage = "https://pypi.python.org/pypi/heavyedge-features/"
source = "https://github.com/heavyedge/heavyedge-features"
documentation = "https://heavyedge-features.readthedocs.io"

[project.optional-dependencies]
test = [
    "pytest",
    "tqdm",
    "heavyedge-classify>=1.1.1,<2.0.0",
]
doc = [
    "sphinx",
    "numpydoc",
    "pydata_sphinx_theme",
]
dev = [
    "flake8",
    "black",
    "isort",
    "heavyedge-features[test,doc]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
doctest_optionflags = [
    "NORMALIZE_WHITESPACE",
    "IGNORE_EXCEPTION_DETAIL",
    "ELLIPSIS",
]
addopts = "--ignore=doc --doctest-modules --import-mode=importlib"

[project.entry-points."heavyedge.commands"]
"Feature quantification" = "heavyedge_features.plugins.features"
